requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pandas>=2.2.3
supabase==2.10.0
lxml==5.1.0
python-dotenv==1.0.0
pyarrow>=11.0.0
sendgrid==6.11.0
